// gestion-de-seguimiento-de-tareas-y-proyectos/src/app/domain/interfaces/member/update-member.interface.domain.ts
export interface IUpdateMemberModel {
    _id:string | null | undefined;
    name: string| null | undefined;
    document: string| null | undefined;
    salary: number| null | undefined;
    role: string| null | undefined;
    email: string| null | undefined;
    password: string | null | undefined;
}